third_party/WebKit/LayoutTests/external/wpt/selection/collapseToStartEnd-expected.txt

CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
CONSOLE ERROR: line 27: The given range isn't in document.
CONSOLE ERROR: line 80: The given range isn't in document.
This is a testharness.js-based test.
Found 118 tests; 58 PASS, 60 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Range 0 [] collapseToStart() 
PASS Range 0 [] collapseToEnd() 
PASS Range 1 [paras[0].firstChild, 0, paras[0].firstChild, 0] collapseToStart() 
PASS Range 1 [paras[0].firstChild, 0, paras[0].firstChild, 0] collapseToEnd() 
PASS Range 2 [paras[0].firstChild, 0, paras[0].firstChild, 1] collapseToStart() 
PASS Range 2 [paras[0].firstChild, 0, paras[0].firstChild, 1] collapseToEnd() 
PASS Range 3 [paras[0].firstChild, 2, paras[0].firstChild, 8] collapseToStart() 
PASS Range 3 [paras[0].firstChild, 2, paras[0].firstChild, 8] collapseToEnd() 
PASS Range 4 [paras[0].firstChild, 2, paras[0].firstChild, 9] collapseToStart() 
PASS Range 4 [paras[0].firstChild, 2, paras[0].firstChild, 9] collapseToEnd() 
PASS Range 5 [paras[1].firstChild, 0, paras[1].firstChild, 0] collapseToStart() 
PASS Range 5 [paras[1].firstChild, 0, paras[1].firstChild, 0] collapseToEnd() 
PASS Range 6 [paras[1].firstChild, 0, paras[1].firstChild, 1] collapseToStart() 
PASS Range 6 [paras[1].firstChild, 0, paras[1].firstChild, 1] collapseToEnd() 
PASS Range 7 [paras[1].firstChild, 2, paras[1].firstChild, 8] collapseToStart() 
PASS Range 7 [paras[1].firstChild, 2, paras[1].firstChild, 8] collapseToEnd() 
PASS Range 8 [paras[1].firstChild, 2, paras[1].firstChild, 9] collapseToStart() 
PASS Range 8 [paras[1].firstChild, 2, paras[1].firstChild, 9] collapseToEnd() 
FAIL Range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 10 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 10 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 11 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 11 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 13 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 13 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 14 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 14 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
PASS Range 15 [document.documentElement, 0, document.documentElement, 1] collapseToStart() 
PASS Range 15 [document.documentElement, 0, document.documentElement, 1] collapseToEnd() 
PASS Range 16 [document.documentElement, 0, document.documentElement, 2] collapseToStart() 
PASS Range 16 [document.documentElement, 0, document.documentElement, 2] collapseToEnd() 
PASS Range 17 [document.documentElement, 1, document.documentElement, 2] collapseToStart() 
PASS Range 17 [document.documentElement, 1, document.documentElement, 2] collapseToEnd() 
PASS Range 18 [document.head, 1, document.head, 1] collapseToStart() 
PASS Range 18 [document.head, 1, document.head, 1] collapseToEnd() 
PASS Range 19 [document.body, 0, document.body, 1] collapseToStart() 
PASS Range 19 [document.body, 0, document.body, 1] collapseToEnd() 
FAIL Range 20 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 20 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 21 [foreignDoc.head, 1, foreignDoc.head, 1] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 21 [foreignDoc.head, 1, foreignDoc.head, 1] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 22 [foreignDoc.body, 0, foreignDoc.body, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 22 [foreignDoc.body, 0, foreignDoc.body, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
PASS Range 23 [paras[0], 0, paras[0], 0] collapseToStart() 
PASS Range 23 [paras[0], 0, paras[0], 0] collapseToEnd() 
PASS Range 24 [paras[0], 0, paras[0], 1] collapseToStart() 
PASS Range 24 [paras[0], 0, paras[0], 1] collapseToEnd() 
FAIL Range 25 [detachedPara1, 0, detachedPara1, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 25 [detachedPara1, 0, detachedPara1, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 26 [detachedPara1, 0, detachedPara1, 1] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 26 [detachedPara1, 0, detachedPara1, 1] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
PASS Range 27 [paras[0].firstChild, 0, paras[1].firstChild, 0] collapseToStart() 
PASS Range 27 [paras[0].firstChild, 0, paras[1].firstChild, 0] collapseToEnd() 
PASS Range 28 [paras[0].firstChild, 0, paras[1].firstChild, 8] collapseToStart() 
PASS Range 28 [paras[0].firstChild, 0, paras[1].firstChild, 8] collapseToEnd() 
PASS Range 29 [paras[0].firstChild, 3, paras[3], 1] collapseToStart() 
PASS Range 29 [paras[0].firstChild, 3, paras[3], 1] collapseToEnd() 
PASS Range 30 [paras[0], 0, paras[0].firstChild, 7] collapseToStart() 
PASS Range 30 [paras[0], 0, paras[0].firstChild, 7] collapseToEnd() 
PASS Range 31 [testDiv, 2, paras[4], 1] collapseToStart() 
PASS Range 31 [testDiv, 2, paras[4], 1] collapseToEnd() 
PASS Range 32 [testDiv, 1, paras[2].firstChild, 5] collapseToStart() 
PASS Range 32 [testDiv, 1, paras[2].firstChild, 5] collapseToEnd() 
PASS Range 33 [document.documentElement, 1, document.body, 0] collapseToStart() 
PASS Range 33 [document.documentElement, 1, document.body, 0] collapseToEnd() 
FAIL Range 34 [foreignDoc.documentElement, 1, foreignDoc.body, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 34 [foreignDoc.documentElement, 1, foreignDoc.body, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
PASS Range 35 [document, 0, document, 1] collapseToStart() 
PASS Range 35 [document, 0, document, 1] collapseToEnd() 
PASS Range 36 [document, 0, document, 2] collapseToStart() 
PASS Range 36 [document, 0, document, 2] collapseToEnd() 
PASS Range 37 [document, 1, document, 2] collapseToStart() 
PASS Range 37 [document, 1, document, 2] collapseToEnd() 
PASS Range 38 [testDiv, 0, comment, 5] collapseToStart() 
PASS Range 38 [testDiv, 0, comment, 5] collapseToEnd() 
PASS Range 39 [paras[2].firstChild, 4, comment, 2] collapseToStart() 
PASS Range 39 [paras[2].firstChild, 4, comment, 2] collapseToEnd() 
PASS Range 40 [paras[3], 1, comment, 8] collapseToStart() 
PASS Range 40 [paras[3], 1, comment, 8] collapseToEnd() 
FAIL Range 41 [foreignDoc, 0, foreignDoc, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 41 [foreignDoc, 0, foreignDoc, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 42 [foreignDoc, 1, foreignComment, 2] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 42 [foreignDoc, 1, foreignComment, 2] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 43 [foreignDoc.body, 0, foreignTextNode, 36] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 43 [foreignDoc.body, 0, foreignTextNode, 36] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 44 [xmlDoc, 0, xmlDoc, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 44 [xmlDoc, 0, xmlDoc, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 45 [xmlDoc, 1, xmlComment, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 45 [xmlDoc, 1, xmlComment, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 46 [detachedTextNode, 0, detachedTextNode, 8] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 46 [detachedTextNode, 0, detachedTextNode, 8] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 47 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 47 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 48 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 48 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 49 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 49 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 50 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 50 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 51 [detachedComment, 3, detachedComment, 4] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 51 [detachedComment, 3, detachedComment, 4] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 52 [detachedComment, 5, detachedComment, 5] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 52 [detachedComment, 5, detachedComment, 5] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 53 [detachedForeignComment, 0, detachedForeignComment, 1] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 53 [detachedForeignComment, 0, detachedForeignComment, 1] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 54 [detachedForeignComment, 4, detachedForeignComment, 4] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 54 [detachedForeignComment, 4, detachedForeignComment, 4] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 55 [detachedXmlComment, 2, detachedXmlComment, 6] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 55 [detachedXmlComment, 2, detachedXmlComment, 6] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 56 [docfrag, 0, docfrag, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 56 [docfrag, 0, docfrag, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 57 [foreignDocfrag, 0, foreignDocfrag, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 57 [foreignDocfrag, 0, foreignDocfrag, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 58 [xmlDocfrag, 0, xmlDocfrag, 0] collapseToStart() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
FAIL Range 58 [xmlDocfrag, 0, xmlDocfrag, 0] collapseToEnd() assert_equals: Sanity check: rangeCount must equal 1 after addRange() expected 1 but got 0
Harness: the test ran to completion.
